+++ app/src/main/java/com/wepindia/pos/BillingHomeDeliveryActivity.java
@@ -6222,10 +6222,10 @@ public class BillingHomeDeliveryActivity extends WepPrinterBaseActivity implemen
                                         tvDiscountAmount.setText(String.format("%.2f", fTotalDiscount));
                                         tvBillNumber.setText(txtReprintBillNo.getText().toString());
 
-                                        tvIGSTValue.setText(String.format("%.2f", cursor.getFloat(cursor.getColumnIndex("IGSTAmount"))));
-                                        tvTaxTotal.setText(String.format("%.2f", cursor.getFloat(cursor.getColumnIndex("CGSTAmount"))));
+                                        tvIGSTValue.setText(String.format("%.2f", cursor.getDouble(cursor.getColumnIndex("IGSTAmount"))));
+                                        tvTaxTotal.setText(String.format("%.2f", cursor.getDouble(cursor.getColumnIndex("CGSTAmount"))));
                                         tvServiceTaxTotal.setText(String.format("%.2f", cursor.getDouble(cursor.getColumnIndex("SGSTAmount"))));
-                                        tvSubTotal.setText(String.format("%.2f", cursor.getFloat(cursor.getColumnIndex("TaxableValue"))));
+                                        tvSubTotal.setText(String.format("%.2f", cursor.getDouble(cursor.getColumnIndex("TaxableValue"))));
                                         tvBillAmount.setText(String.format("%.2f", cursor.getDouble(cursor.getColumnIndex("BillAmount"))));
                                         edtCustId.setText(cursor.getString(cursor.getColumnIndex("CustId")));
